backend/pyproject.toml
[tool.poetry]
name = "dorapulse"
version = "0.1.0"
description = "This package is for Dora metrics"
authors = ["yellowdraggon <chandan.rattan@gmail.com>"]
license = "Apache 2"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.100.0"}
uvicorn = "^0.23.0"
httpx = "^0.24.0"
python-dotenv = "^1.0.0"
pandas = "^2.0.0"
openpyxl = "^3.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pre-commit = "^3.3.2"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
